test: move field length reviewer tests to new test class
[#OCD-2635]
