+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/validation/listing/reviewer/ValidDataReviewerTest.java
@@ -3,11 +3,6 @@ package gov.healthit.chpl.validation.listing.reviewer;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import java.awt.List;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.Set;
-
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -17,21 +12,16 @@ import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import org.mockito.Spy;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.MessageSource;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import gov.healthit.chpl.domain.CertificationResult;
 import gov.healthit.chpl.domain.CertificationResultAdditionalSoftware;
 import gov.healthit.chpl.domain.CertifiedProductSearchDetails;
-import gov.healthit.chpl.domain.CertifiedProductSed;
-import gov.healthit.chpl.domain.TestParticipant;
-import gov.healthit.chpl.domain.TestTask;
 import gov.healthit.chpl.domain.concept.PrivacyAndSecurityFrameworkConcept;
 import gov.healthit.chpl.exception.EntityRetrievalException;
 import gov.healthit.chpl.listing.ListingMockUtil;
 import gov.healthit.chpl.manager.CertifiedProductManager;
-import gov.healthit.chpl.util.ErrorMessageUtil;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = { gov.healthit.chpl.CHPLTestConfig.class })
@@ -43,18 +33,12 @@ public class ValidDataReviewerTest {
             + PrivacyAndSecurityFrameworkConcept.getFormattedValues();
     private static final String BAD_ADDL_SOFTWARE_ERROR =
             "No CHPL product was found matching additional software CHP-12345 for " + D_1;
-    
-    private static final String BAD_LENGTH_PARTICIPANT_ID = "You have exceeded the max length, 20 characters, for the Participant Identifier ID This is more than twenty characters long..";
-    private static final String BAD_LENGTH_TASK_ID = "You have exceeded the max length, 20 characters, for the Task Identifier with ID This is more than twenty characters long..";
 
     @Spy private CertifiedProductManager cpManager;
     @Autowired private ListingMockUtil mockUtil;
 
     @InjectMocks
     private ValidDataReviewer validDataReivewer;
-    
-    @Autowired
-    private FieldLengthReviewer fieldLengthReivewer;
 
     @Before
     public void setup() {
@@ -132,40 +116,5 @@ public class ValidDataReviewerTest {
         validDataReivewer.review(listing);
         assertFalse(listing.getErrorMessages().contains(BAD_PRIVACY_SECURITY_ERROR));
         assertTrue(listing.getErrorMessages().contains(BAD_ADDL_SOFTWARE_ERROR));
-    }
-    
-    @Test
-    public void testBadTestTask_HasErrors() {
-        CertifiedProductSearchDetails listing = mockUtil.createValid2015Listing();
-        CertifiedProductSed sed = new CertifiedProductSed();
-        TestTask tt = new TestTask();
-        tt.setUniqueId("This is more than twenty characters long.");
-        ArrayList<TestTask> tts = new ArrayList<TestTask>();
-        tts.add(tt);
-        sed.setTestTasks(tts);
-        listing.setSed(sed);
-        fieldLengthReivewer.review(listing);
-        System.out.println(listing.getErrorMessages());
-        assertTrue(listing.getErrorMessages().contains(BAD_LENGTH_TASK_ID));
-    }
-    
-    @Test
-    public void testBadTestParticipant_HasErrors() {
-        CertifiedProductSearchDetails listing = mockUtil.createValid2015Listing();
-        CertifiedProductSed sed = new CertifiedProductSed();
-        TestTask tt = new TestTask();
-        tt.setUniqueId("participant");
-        ArrayList<TestTask> tts = new ArrayList<TestTask>();
-        tts.add(tt);
-        sed.setTestTasks(tts);
-        TestParticipant tp = new TestParticipant();
-        tp.setUniqueId("This is more than twenty characters long.");
-        Set<TestParticipant> tps = new HashSet<TestParticipant>();
-        tps.add(tp);
-        sed.getTestTasks().get(0).setTestParticipants(tps);
-        listing.setSed(sed);
-        fieldLengthReivewer.review(listing);
-        System.out.println(listing.getErrorMessages());
-        assertTrue(listing.getErrorMessages().contains(BAD_LENGTH_PARTICIPANT_ID));
     }
 }
